commit code after filter code and resolve bugs

+++ lib/features/personal/personal_profile/view/self_employed/model/earn_service_model_response.dart
@@ -408,8 +408,8 @@ class SocialLinks {
 }
 
 class UserLocation {
-  int? lat;
-  int? lon;
+  double? lat;
+  double? lon;
 
   UserLocation({this.lat, this.lon});
